refactor(mobile): scope measurement-flow-screen subscription and extract hooks
The screen pulled 9 fields out of useMeasurementFlowStore with a plain
object destructure, which subscribes to the whole store and re-renders
on every flow tick (scanResult, currentFlowStep, iterationCount …). That
cascade hit MeasurementFlowContainer and its children every time and
was a real factor in the click-then-Start lag on slower phones.

- Subscribe only to experimentId via a selector; everything else used
  by the auto-pause effect is read via getState() when the effect
  fires (i.e. on blur), which is the only moment that data is needed.
- Extract useAutoPauseFlow and useFlowBackHandler so the screen body
  isn't a tangle of effects and the hook deps are minimal (isFocused,
  experimentLabel) instead of the previous 10-item list.
- Replace JSON.parse(JSON.stringify(answersHistory)) with the native
  structuredClone — same detach-from-live-store guarantee, faster.
- Read the title from useExperiment(experimentId) instead of running
  the whole useExperiments list query and mapping/find-ing it on every
  render just to pull one label.
